bookmark: Add `advance`, a specialized move with `--to`/`--from` defaults
A common recommendation for new users is for them to add a `jj tug`
alias. This change makes that alias unnecessary by adding a native
version.

This change adds `close_bookmarks(to)` and `closest_pushable(to)` revsets
which act as customization points for users who need more specific behavior.

The default values of the revsets are inspired by the discussion in
https://github.com/jj-vcs/jj/discussions/5568#discussioncomment-12674748.

Co-authored-by: Scott Taylor <scott11x8@gmail.com>